test(attn): VT_KV_CACHE_F32 runtime toggle for same-binary KV-dtype A/B
full_attn_buf_ is now a raw byte buffer sized by the chosen KV-cache dtype
(bf16 default, f32 when VT_KV_CACHE_F32=1), enabling a clean f32-vs-bf16
throughput/RAM comparison in one binary.

+++ include/vllm/v1/worker/gpu/runner.h
@@ -207,11 +207,12 @@ class GPUModelRunner final : public ModelRunnerBase {
   std::vector<int32_t> gdn_free_slots_;
 
   // Owned KV-cache backing storage (host at T0) + the views the forward reads.
-  // full_attn_buf_ is bf16 (uint16_t words) — the paged KV cache stores K/V in
-  // bf16, mirroring vLLM's bf16 flash_attn KV store (halves KV memory).
-  std::vector<std::vector<uint16_t>> full_attn_buf_;  // per full-attn layer (bf16)
-  std::vector<std::vector<float>> ssm_buf_;           // per GDN layer
-  std::vector<std::vector<float>> conv_buf_;          // per GDN layer
+  // full_attn_buf_ is a raw byte buffer sized by the KV-cache dtype: bf16 by
+  // default (mirrors vLLM's bf16 flash_attn KV store, halves KV memory), or f32
+  // when VT_KV_CACHE_F32 is set (same-binary A/B).
+  std::vector<std::vector<uint8_t>> full_attn_buf_;  // per full-attn layer (bytes)
+  std::vector<std::vector<float>> ssm_buf_;          // per GDN layer
+  std::vector<std::vector<float>> conv_buf_;         // per GDN layer
   std::vector<PagedKvCache> attn_kv_;
   std::vector<GdnStateCache> gdn_state_;
